交通部分重新上傳

+++ src/innocence741/model/HighSpeedRail.java
@@ -1,5 +1,6 @@
 package innocence741.model;
 
+import java.math.BigDecimal;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -19,7 +20,7 @@ public class HighSpeedRail {
 	@Id
 	@Column(name = "SN_SCHEDULE")
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int snSchedule;
+	private BigDecimal snSchedule;
 	
 	@Column(name = "ID_HSR")
 	private String idHSR;
@@ -67,13 +68,13 @@ public class HighSpeedRail {
 	private Set<T_Order_List> t_Order_Lists = new HashSet<T_Order_List>();
 
 	
-
-
-	public int getSnSchedule() {
+	
+	
+	public BigDecimal getSnSchedule() {
 		return snSchedule;
 	}
 
-	public void setSnSchedule(int snSchedule) {
+	public void setSnSchedule(BigDecimal snSchedule) {
 		this.snSchedule = snSchedule;
 	}
 
@@ -188,7 +189,7 @@ public class HighSpeedRail {
 	public void setZuoying(String zuoying) {
 		this.zuoying = zuoying;
 	}
-	
+
 	public Set<T_Order_List> getT_Order_Lists() {
 		return t_Order_Lists;
 	}
@@ -196,9 +197,7 @@ public class HighSpeedRail {
 	public void setT_Order_Lists(Set<T_Order_List> t_Order_Lists) {
 		this.t_Order_Lists = t_Order_Lists;
 	}
-	
-	
-	
+
 	public String getArriveTime(String location) {
 		if(location.equals("Nangang")) {
 			return getNangang();
